feat: referral system, improve miniapp support, fixes
- added referral system settings page
- implemented draft referral system structure (models, service, repository)
- implemented hash generator for referral links
- added default values for environment variables
- renamed BOT_MINI_APP_URL to BOT_MINI_APP and updated its behavior
- added support for opening the subscription page in WebApp (for legiz/Orion)
- improved panel information display
- improved statistics display
- fixed incorrect end date when importing an expired user
- corrected translations
- updated README and env.example

+++ src/bot/middlewares/garbage.py
@@ -22,7 +22,7 @@ class GarbageMiddleware(EventTypedMiddleware):
         message = cast(Message, event)
         user: UserDto = data[USER_KEY]
 
-        if message.text != f"/{Command.START.value.command}":
+        if message.text and not message.text.startswith(f"/{Command.START.value.command}"):
             await message.delete()
             logger.debug(f"Message '{message.content_type}' deleted from '{user.telegram_id}'")
